tests/wpt/web-platform-tests/css/css-fonts/parsing/font-synthesis-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: parsing font-synthesis with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-synthesis-prop">
<meta name="assert" content="font-synthesis supports only the grammar 'none | [ weight || style ]'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value('font-synthesis', 'auto');
test_invalid_value('font-synthesis', 'none weight');
</script>
</body>
</html>
